% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{alignMatrices}
\alias{alignMatrices}
\title{Align adjacency matrices}
\usage{
alignMatrices(A)
}
\arguments{
\item{A}{A list of two square (weighted) adjacency matrices.}
}
\value{
The output of \code{alignMatrices()} follows the same format as its
  input. It produces two adjacency matrices in a list. Each matrix
  corresponds to the respective input entry. The input matrices have been
  reduced to their common row and column labels and reordered so their
  indeces match.
}
\description{
\code{alignMatrices()} reduces two adjacency matrices to their common nodes
and ordering them to match. The matrices can then be compared using different
similarity measures. Note alignment is not always necessary, since many
network construction algorithms will preserve the same node order across
different runs.
}
\examples{
# Create two symmetric matrices of different sizes with overlapping labels.
A1 <- matrix(runif(16), nrow=4, ncol=4); A1 <- A1 + t(A1)
rownames(A1) <- colnames(A1) <- LETTERS[1:4]
A2 <- matrix(runif(25), nrow=5, ncol=5); A2 <- A2 + t(A2)
rownames(A2) <- colnames(A2) <- LETTERS[2:6]
# Align them. This will result in two 3x3 matrices with row and
# column labels B, C, D.
alignMatrices(list(A1, A2))
}
