% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{calculateKSimilarity}
\alias{calculateKSimilarity}
\title{Rank k-similarity}
\usage{
calculateKSimilarity(D1, D2, k.or.p = 0.1)
}
\arguments{
\item{D1, D2}{Numeric arrays of equal length and labels, such as node metric
values (e.g. two alternative degree sequences).}

\item{k.or.p}{Either an integer (k) or a fractiion(p). The number of highest
ranking nodes to compare}
}
\value{
The rank k-similarity of two aligned vectors is the overlap of the
  top k highest ranking values across them. Rank k-similarity is given as a
  proportion.

  For example, if the three highest ranking nodes in one network are A, B,
  and C, and in another they are B, C, and D, the overlap between the top
  three highest ranking nodes is of size 2. The rank k-similarity between the
  degree sequences for k=3 is therefore 2/3. If a proportion is given, e.g.
  \code{k.or.p=0.1}, the top \code{k.or.p*length(D1)} nodes are considered
  for the overlap.
}
\description{
Rank k-similarity is the overlap between the highest ranking entries of two
numeric arrays. It is useful in network analysis, where two aligned networks
migh be considered similar if they share key nodes.
}
\examples{
# Generate two node metric vectors
D1 <- rpois(100, 4); D2 <- rpois(100, 3)
# Calculate their rank k-similarity for the top k=10 nodes.
calculateKSimilarity(D1, D2, k.or.p=10)
# Calculate their rank k-similarity for the top p=15\% of nodes.
calculateKSimilarity(D1, D2, k.or.p=0.15)

}
\references{
Bozhilova, Lyuba V., et al. "Measuring rank robustness in scored
  protein interaction networks." BMC Bioinformatics 20.1 (2019): 1-14.

  Trajanovski, Stojan, et al. "Robustness envelopes of networks." Journal of
  Complex Networks 1.1 (2013): 44-62.
}
