% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputCheck.R
\name{checkArrayList}
\alias{checkArrayList}
\title{Check node metric list}
\usage{
checkArrayList(D)
}
\arguments{
\item{D}{A list of two numeric arrays, corresponding to node metrics.}
}
\value{
If the input passes the check, the output is \code{TRUE}. Otherwise,
  one of two errors is thrown, corresponding to each of the two checks
  performed by \code{checkArrayList()}. These checks are:
  \itemize{
  \item Is \code{D} a list of length 2?
  \item Is there a label mismatch? Either both
  arrays or neither array should be named.
  }
}
\description{
\code{checkArrayList()} checks whether its input is a list of arrays.
}
\examples{
D1 <- runif(10); names(D1) <- LETTERS[1:10]
D2 <- runif(10); names(D2) <- LETTERS[5:14]
checkArrayList(list(D1, D2))

}
