% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputCheck.R
\name{checkExpressionDF}
\alias{checkExpressionDF}
\title{Check a gene expression data frame}
\usage{
checkExpressionDF(df)
}
\arguments{
\item{df}{A COGENT-compatible gene expression data frame.}
}
\value{
If the input passes the check, the output is \code{TRUE}. Otherwise,
  one of three errors is thrown, corresponding to each of the three checks
  performed by \code{checkExpressionDF()}. These checks are:
  \itemize{
  \item Is \code{df} a data frame?
  \item Does it have a column \code{Name} where gene names are recorded?
  \item Are all other columns of type numeric?
  }
  Note the \code{Name} column is not required to be of type \code{character}.
  Numeric gene indeces, \code{NA}s and \code{NULL}s are allowed.
}
\description{
\code{checkExpressionDF()} checks whether its input is a COGENT-compatible
gene expressiom data frame.
}
\examples{
df <- as.data.frame(matrix(runif(500)), nrow=10, ncol=50)
df <- cbind(Name=LETTERS[1:10], df)
checkExpressionDF(df)

}
