% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputCheck.R
\name{checkMatrixList}
\alias{checkMatrixList}
\title{Check adjacency matrix list}
\usage{
checkMatrixList(A)
}
\arguments{
\item{A}{A list of two square (weighted) adjacency matrices.}
}
\value{
If the input passes the check, the output is \code{TRUE}. Otherwise,
  one of three errors is thrown, corresponding to each of the three checks
  performed by \code{checkMatrixList()}. These checks are:
  \itemize{
  \item Is \code{A} a list of length 2?
  \item Are the matrices it contains square?
  \item Is there a label mismatch? Column names should be either present in
  both or absent in both matrices.
  }
}
\description{
\code{checkMatrixList()} checks whether its input is a list of matrices of
length two. It does \strong{not} check whether the matrices are symmetric.
}
\examples{
expressionData <- matrix(runif(25*20), nrow=25, ncol=20)
A1 <- cor(expressionData[,1:10]); A2 <- cor(expressionData[,11:20])
checkMatrixList(list(A1, A2))

}
