% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkSimilarity.R
\name{getEdgeSimilarity}
\alias{getEdgeSimilarity}
\title{Get the edge similarity for two networks}
\usage{
getEdgeSimilarity(A, align = FALSE)
}
\arguments{
\item{A}{A list of two square (weighted) adjacency matrices.}

\item{align}{Logical; Whether to align the two adjacency matrices. Only set
to TRUE if this is not done automatically.}
}
\value{
The result is a list of similarity measures of the two networks with
  adjacency matrices described by \code{A}. This includes:
  \itemize{
  \item \code{nodeCount} The number of non-isolated genes across the two
  networks.
  \item \code{globalSimilarity} The weighted Jaccard index of the edge sets
  of the two networks.
  \item \code{localSimilarity} The weighted Jaccard index for
  each gene neighbourhood.
  }
}
\description{
\code{getEdgeSimilarity()} calculates the weighted or unweighted Jaccard
index between the edge sets of two networks. The calculation is performed
both for the full edge set and for each node neighbourhood.
}
\examples{
# Generate two adjacency matrices
A1 <- matrix(0, ncol=10, nrow=10); A2 <- matrix(0, ncol=10, nrow=10)
A1[upper.tri(A1)] <- rbinom(45, 1, .2); A1 <- A1+t(A1)
A2[upper.tri(A1)] <- rbinom(45, 1, .4); A2 <- A2+t(A2)
colnames(A1) <- rownames(A1) <- LETTERS[1:10]
colnames(A2) <- rownames(A2) <- LETTERS[6:15]
# Calculate similarity
getEdgeSimilarity(list(A1, A2), align=TRUE)

}
