% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{splitExpressionData}
\alias{splitExpressionData}
\title{Randomly subset expression samples}
\usage{
splitExpressionData(df, propShared = 0)
}
\arguments{
\item{df}{A COGENT-compatible data frame with rows corresponding to genes and
columns corresponding to samples. A column called Name containing gene
names is expected.}

\item{propShared}{The proportion of the samples in \code{df} shared across
the two subsets.}
}
\value{
The function returns a list containing two COGENT-compatible data
  frames. Every sample of the original data frame belongs to at least one of
  the two data frames in the outputs. A total of propShared*numSamples are
  shared across the two. If the total number of samples is odd, a single
  sample will be omitted at random to ensure the equal sizing of the split.
}
\description{
This function takes a COGENT-compatible expression data frame and divides its
samples in two equally sized groups. Sample overlap between the groups is
allowed.
}
\examples{
# Generate a COGENT-compatible data frame
df <- as.data.frame(matrix(runif(500), nrow=10, ncol=50))
df <- cbind(Name=LETTERS[1:10], df)
# Split with no sample overlap
splitExpressionData(df, propShared=0)
# Split so 50\% of the samples are shared, and 25\% each are unique to
# one of the two subsets.
splitExpressionData(df, propShared=0.50)

}
